Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Nice, reboot to recovery now works.

Not so nice: using CWM recovery, I need to adb in, mount /system, and call

/system/bin/mischelp /dev/block/platform/sdhci-tegra.3/by-name/MC

or the system will boot to CWM always.

That will surprise some users, I guess...

If the mechanism cannot be changed to be one-shot by itself, maybe it would be prudent and possible to have mischelp (when triggered in the running system) to look into the installed recovery, to test whether that recovery appears to be set up to undo the switch (look for /sbin/mischelp in there, and/or the calling entry in init.rc)

Patch-set: 1