Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Maybe we should also update to stock keylayout. 

I'm not entirely sure about that, as the valid solution should be to load a keymap based on the language setting automatically (and allow config too as done in AutoParts).

Patch-set: 1