Update patch set 1
Patch Set 1:

At the risk of showing my ignorance - and with all due respect to the work that certainy went into this...:

Why is this based on 3.1, a kernel series explicitly EOLed since January? What are the plans for the future?

I like the situation with GNex - 3.0.x kernels, declared long-term supported by GregKH, and Rafael merging upstream updates within a few days.

Patch-set: 1