Update overlay for fixed lid logic
Use correct value for lidKeyboardAccessibility to hide the software
keyboard if docked.

Also remove lidOpenRotation. Though forcing landscape orientation if
dock is attached would be nice, the framework doesn't know about
removeable lids at all. So if the dock is removed it still applies the
lidOpenRotation value, because essentially the lid is still open.

This is a temporary workaround. I plan on adding a config value that
tells the framework to ignore lidOpenRotation if the dock is absent, but
that patch is still in the works.

Change-Id: Ifd93133a26f7e9976997ad1b9d9d8b987a093188